requirements.txt
MetaTrader5>=5.0.45
pandas>=2.0.0
numpy>=1.24.0
scikit-learn>=1.3.0
xgboost>=2.0.0
joblib>=1.3.0
requests>=2.31.0
python-dateutil>=2.8.2
pytz>=2023.3
